Add files via upload
This application is a simple and totally complete book library system, It helps Jeddah
university students to use library services easily.
Here in this project, the student can book a place to study or Borrowing a book from
the books that available in the library. Also, the student can book to use
Photocopying / Printing Service.